requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt==4.3.0
sqlalchemy
pydantic
pydantic[email]
pydantic-settings
python-dotenv
psycopg2-binary
python-multipart